+++ machines/devbox/configuration.nix
@@ -20,32 +20,9 @@
   # Enable networking
   networking.networkmanager.enable = true;
 
-  # internationalization.
-  i18n.defaultLocale = "en_DK.UTF-8";
-
-  i18n.extraLocaleSettings = {
-    LC_ADDRESS = "da_DK.UTF-8";
-    LC_IDENTIFICATION = "da_DK.UTF-8";
-    LC_MEASUREMENT = "da_DK.UTF-8";
-    LC_MONETARY = "da_DK.UTF-8";
-    LC_NAME = "da_DK.UTF-8";
-    LC_NUMERIC = "da_DK.UTF-8";
-    LC_PAPER = "da_DK.UTF-8";
-    LC_TELEPHONE = "da_DK.UTF-8";
-    LC_TIME = "da_DK.UTF-8";
-  };
-
   # Users.
-  users.users.daniel = {
-    extraGroups = [
-      "networkmanager"
-      "wheel"
-      "vboxsf"
-    ];
-  };
-  users.users.caddy = {
-    extraGroups = [ "vboxsf" ];
-  };
+  users.users.daniel.extraGroups = [ "vboxsf" ];
+  users.users.caddy.extraGroups = [ "vboxsf" ];
 
   # System specific software.
   environment.systemPackages = with pkgs; [